requirements.txt
python-telegram-bot>=21.0
yt-dlp==2024.3.10
instaloader==4.11
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
pillow==11.0.0
streamlit>=1.28.0
psutil>=5.9.0
